// src/front/app/dto/search-constraints-request.interface.ts
/*
  AUTOGENERATED FILE! DO NOT EDIT!
  GENERATED BY 'DTORecordToAngularInterfaceConverter.java'
  FROM THE JAVA SIDE OF THIS PROJECT
*/
import { FileSearchConstraints } from './file-search-constraints.interface';

export interface SearchConstraintsRequest {
    fileConstraints: FileSearchConstraints;
}