// src/front/app/dto/file-search-constraints.interface.ts
/*
  AUTOGENERATED FILE! DO NOT EDIT!
  GENERATED BY 'DTORecordToAngularInterfaceConverter.java'
  FROM THE JAVA SIDE OF THIS PROJECT
*/
import { SearchConstraintCondition } from './search-constraint-condition.enum';
import { SearchConstraintRange } from './search-constraint-range.interface';

export interface FileSearchConstraints {
    directory: SearchConstraintCondition;
    hidden: SearchConstraintCondition;
    link: SearchConstraintCondition;
    special: SearchConstraintCondition;
    date: SearchConstraintRange;
    size: SearchConstraintRange;
    storages: Array<string>;
    parentPath: string;
    parentHashPath: string;
}